+++ packages/react-workspace/nimble-react/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-react
 
-<!-- This log was last generated on Mon, 02 Mar 2026 00:16:21 GMT and should not be manually modified. -->
+<!-- This log was last generated on Mon, 02 Mar 2026 17:43:16 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.11.10
+
+Mon, 02 Mar 2026 17:43:16 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.3.4
 
 ## 0.11.9